refactor: modularize DirectClient API router and improve code organization
- Extract message handler template to a dedicated `templates.ts` file
- Create `types.ts` to define CustomRequest interface for file uploads
- Move API router logic from `index.ts` to `api.ts` for better separation of concerns
- Simplify DirectClient initialization by removing inline route handlers
- Improve code readability and maintainability by breaking down complex methods
