+++ packages/client-direct/src/types.ts
@@ -1,0 +1,5 @@
+import type { Request as ExpressRequest } from "express";
+
+export interface CustomRequest extends ExpressRequest {
+    file?: Express.Multer.File;
+}
